V0.1 to asset store

+++ Assets/Scripts/UI/PlayerShieldUI.cs
@@ -22,8 +22,8 @@ namespace UI
 			if (player.IsShieldUnlocked())
 			{
 				CheckNeedToFlash(currentHealth, maxHealth);
-				tmp.text = Utility.FormatMoneyToKMB( currentHealth) + "/" +
-				           Utility.FormatMoneyToKMB( maxHealth);
+				tmp.text = Utility.FormatMoneyToKMB( (long)currentHealth) + "/" +
+				           Utility.FormatMoneyToKMB((long) maxHealth);
 				icon.enabled = true;
 				icon.color = defaultImageColor;
 				if (currentHealth != 0 || cor == null) return;
